test(logic): split economy stockpile preview part2 tests for #2288
Replace the monolithic economy_stockpile_preview_part2_test.dart (884 lines)
with three segment files under the 400-line incremental gate, preserving
the same assertions and grouping semantics.

Refs #2288
